Update: ArticleItem Shadow DOM 삭제 및 formatDate 분리
