+++ src/components/ArticleItem/ArticleItem.js
@@ -1,13 +1,14 @@
 import template from "./template.html";
-import style from "./style.css?shadow";
+import "./style.css";
+
+import { formatDate } from "../../utils/formatDate";
 
 export default class ArticleItem extends HTMLElement {
   constructor() {
     super();
-
-    if (!this.shadowRoot) {
-      this.attachShadow({ mode: "open" });
-    }
+    this.template = new DOMParser()
+      .parseFromString(template, "text/html")
+      .querySelector("template").content;
 
     this.data = {};
   }
@@ -17,33 +18,22 @@ export default class ArticleItem extends HTMLElement {
     this.render();
   }
 
-  formatDate(dateString) {
-    const date = new Date(dateString);
-    const year = date.getFullYear();
-    const month = date.getMonth() + 1;
-    const day = date.getDate();
-
-    return `${year}. ${month}. ${day}`;
-  }
-
   render() {
-    const HTMLTemplate = new DOMParser()
-      .parseFromString(template, "text/html")
-      .querySelector("template").content;
-    const styleElement = document.createElement("style");
-    styleElement.textContent = style;
-
-    HTMLTemplate.querySelector(".article-item__a").href =
+    this.template.querySelector(".article-item__a").href =
       "article/" + this.data.id;
-    HTMLTemplate.querySelector(".article-item__image").src = this.data.imageUrl;
-    HTMLTemplate.querySelector(".article-item__title").textContent =
+    this.template.querySelector(".article-item__image").src =
+      this.data.imageUrl;
+    this.template.querySelector(".article-item__title").textContent =
       this.data.title;
-    HTMLTemplate.querySelector(".article-item__description").textContent =
+    this.template.querySelector(".article-item__description").textContent =
       this.data.description;
-    HTMLTemplate.querySelector(".article-item__date").textContent =
-      this.formatDate(this.data.createdAt);
+    this.template.querySelector(".article-item__date").textContent = formatDate(
+      this.data.createdAt
+    );
 
-    this.shadowRoot.appendChild(HTMLTemplate.cloneNode(true));
-    this.shadowRoot.appendChild(styleElement);
+    const liElement = document.createElement("li");
+    liElement.appendChild(this.template.cloneNode(true));
+
+    this.appendChild(liElement);
   }
 }
